msm: sde: fix wrong chroma stride for h2v1/h1v2 formats
Current driver is swapping the chroma subsample
for the read pipe when rotation is in place,
this configuration is already taken care by the
rotator, and it is not needed to swap the chroma
sampling for the rotator pipes.
This fixes wrong stride calculations, causing
wrong colors shown in h2v1/h1v2 tests.

CRs-Fixed: 1069748
Change-Id: Idfbe229a39bda728aebbdecd45e468b808026ff8
Signed-off-by: Alan Kwong <akwong@codeaurora.org>