msm: camera: Use correct direct field size while copying
In the payload.direct field of abort and destroy ICP commands,
the size was previously 4 bytes. But ICP firmware expects this
to be 8 bytes. Fix this to the correct size of 8 bytes.
Allocate the payload dynamically now rather than on the stack.
This will prevent unnecessary allocation on the stack because
the payload size may vary.

Change-Id: Idc6b49c16784399686982ccc2bdc4118b40b666f
Signed-off-by: Karthik Anantha Ram <kartanan@codeaurora.org>