msm: vidc: HDR10 PQ SEI generation for HEVC encoder
To enable HDR10 PQ feature, the userspace sets controls
for content light level and mastering display info. A
single property is set for both the controls to video
hardware so as to generate the SEI in encoded bitstream.
Use s64 datatype for minimum, maximum and default control
value to support UINT_MAX.

CRs-Fixed: 2100854
Change-Id: I42f8b11c5d3ccf0457d8088b2bdb30de3744e211
Signed-off-by: Umesh Pandey <umeshp@codeaurora.org>