msm: camera: icp: Populate the right size for pkts sent by FW
FW response packtets include payload field in their response
structs. The size of this field is variable we need to obtain
the right size using the size field and not calculate the static
size alone. This needs to be done for all response packets from
FW eventhough icp kmd does not process certain packets.

Change-Id: Ie9b1ea3b524e6745d7508632cb3f5605b058bb2b
Signed-off-by: Karthik Anantha Ram <kartanan@codeaurora.org>